Refactors to Metrics::Algorithms
Moves both types of metrics on algorithms to use a single class and
associated database table.

Creating the namespace `Metrics` should allow us to better organize code
associated with various metrics into a common space that is both
intuitive and useful.

Relevant ticket(s):
https://mitlibraries.atlassian.net/browse/TCO-17
