--- test/fixtures/aggregate_matches.yml
@@ -1,23 +1,0 @@
-# == Schema Information
-#
-# Table name: aggregate_matches
-#
-#  id         :integer          not null, primary key
-#  doi        :integer
-#  issn       :integer
-#  isbn       :integer
-#  pmid       :integer
-#  unmatched  :integer
-#  created_at :datetime         not null
-#  updated_at :datetime         not null
-#
-
-# This model initially had no columns defined. If you add columns to the
-# model remove the "{}" from the fixture names and add the columns immediately
-# below each fixture, per the syntax in the comments below
-#
-one: {}
-# column: value
-#
-two: {}
-# column: value
